refactor: ディレクトリ構造を修正
