+++ next-app/src/features/error/AuthError.ts
