+++ next-app/src/features/error/DbError.ts
