+++ next-app/src/features/error/ResponseError.ts
