+++ next-app/src/features/error/SecurityError.ts
