+++ next-app/src/features/error/UserError.ts
