+++ next-app/src/features/error/index.ts
