+++ next-app/src/features/projects/repository/index.ts
@@ -1,0 +1,1 @@
+export { ProjectsRepository } from "./project.repository"
